+++ src/plugins/index.js
@@ -19,6 +19,7 @@ import { followUpRemindersPlugin } from './examples/followUpRemindersPlugin.jsx'
 import { gamificationPlugin } from './examples/gamificationPlugin.jsx';
 import { paperReferenceProspectorPlugin } from './examples/paperReferenceProspectorPlugin.jsx';
 import { searchPrefixProviderPlugin } from './examples/searchPrefixProviderPlugin.jsx';
+import { searchDSLProviderPlugin } from './examples/searchDSLProviderPlugin.jsx';
 import { graphLinterPlugin } from './examples/graphLinterPlugin.jsx';
 import { centralityLensPlugin } from './examples/centralityLensPlugin.jsx';
 
@@ -47,6 +48,7 @@ export const bundledCustomPlugins = [
   gamificationPlugin,
   paperReferenceProspectorPlugin,
   searchPrefixProviderPlugin,
+  searchDSLProviderPlugin,
   graphLinterPlugin,
   centralityLensPlugin,
 ];